chore: relocate component
